+++ src/Agent/NewRelic/Agent/Core/DataTransport/ConnectionHandler.cs
@@ -426,7 +426,7 @@ namespace NewRelic.Agent.Core.DataTransport
                     var responseBody = wire.SendData(method, _connectionInfo, serializedData, requestGuid);
                     return ParseResponse<T>(responseBody);
                 }
-                catch (Exceptions.HttpException ex)
+                catch (DataTransport.HttpException ex)
                 {
                     Log.DebugFormat("Request({0}): Received a {1} {2} response invoking method \"{3}\" with payload \"{4}\"", requestGuid, (int)ex.StatusCode, ex.StatusCode, method, serializedData);
