+++ src/Agent/NewRelic/Agent/Core/DataTransport/HttpException.cs
@@ -4,7 +4,7 @@
 using System;
 using System.Net;
 
-namespace NewRelic.Agent.Core.Exceptions
+namespace NewRelic.Agent.Core.DataTransport
 {
     /// <summary>
     /// Thrown when the connection to the collector(RPM) reports an HTTP transport error.
